backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
sqlalchemy
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv
joblib==1.3.2
numpy==1.24.3
python-multipart==0.0.6
aiofiles==23.2.1
fastapi-cors
xgboost
shap
groq

